+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHKJDHR47FEQ0K5QETGSDG04"
+        "01KHN2X344TSB3CG78WCQASHED"
     )
     r"""The unique identifier of the data source"""