+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHKJDHQSG4H519Z02MACZ07Q"
+        "01KHN2X34127A9HHVFY1QH3TBF"
     )
     r"""The unique identifier of the data source"""