+++ packages/orq-rc/RELEASES.md
@@ -5719,3 +5719,13 @@ Based on:
 - [python v4.4.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.3 - packages/orq-rc
+
+## 2026-02-17 06:02:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.720.0 (2.827.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.4] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.4 - packages/orq-rc